Url Fuzzing testing
The new files change the program from a simple GUI to a similarly crappy GUI that actually does things. It takes a URL input, then tests that input for GET SQL and XSS attack vulnerabilities, then displays which vulnerabilities exist if any.
:::::::::::CHANGES:::::::::::
Added a text-box (Editable / non-scrollable). Program takes values in the text-box and performs several HTTP requests.
Label now outputs which variables are vulnerable and which kinds of attack they are vulnerable to.
When button is pressed, the value in textbox(must be request with php GET values) is cleansed down to its individual variables, then each will be given an illegal XSS or SQLI  addition, then a new HTTP request will be issued with fuzzed data, and any valid responses are displayed in the programs label as vulnerable to the tested addition.
Added new log entry to development log.